% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtm_train_GLMNET_MAXENT_SVM.R
\docType{data}
\name{dtm_train_GLMNET_MAXENT_SVM}
\alias{dtm_train_GLMNET_MAXENT_SVM}
\title{DTM used in the GLMNET_MAXENT_SVM model}
\format{An object of class \code{DocumentTermMatrix} (inherits from \code{simple_triplet_matrix}) with 60118 rows and 258 columns.}
\usage{
dtm_train_GLMNET_MAXENT_SVM
}
\description{
This object...
}
\keyword{datasets}
