% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtm_train_NB.R
\docType{data}
\name{dtm_train_NB}
\alias{dtm_train_NB}
\title{DTM used in the NB model}
\format{An object of class \code{DocumentTermMatrix} (inherits from \code{simple_triplet_matrix}) with 60118 rows and 258 columns.}
\usage{
dtm_train_NB
}
\description{
This object...
}
\keyword{datasets}
