% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofaaasDataset.R
\docType{data}
\name{gofaaasDataset}
\alias{gofaaasDataset}
\title{The gofaaas Dataset}
\format{An object of class \code{data.frame} with 511 rows and 5 columns.}
\usage{
gofaaasDataset
}
\description{
This object contains the text of a news article...
}
\keyword{datasets}
