% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_XGB_vfn.R
\docType{data}
\name{model_XGB_vfn}
\alias{model_XGB_vfn}
\title{XGBoost Model}
\format{An object of class \code{xgb.Booster} of length 9.}
\usage{
model_XGB_vfn
}
\description{
This object...
}
\keyword{datasets}
