% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelNB.R
\name{modelNB}
\alias{modelNB}
\title{modelNB Function}
\usage{
modelNB(fileName)
}
\arguments{
\item{fileName}{}
}
\description{
This function allows you to check if a news article is authentic or fake.
It was trained using the Naive Bayes Algorithms from the package e1071, and the the Kaggle Fake News Dataset
}
\examples{
modelNB(FileName)
}
