% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_GLMNET_MAXENT_SVM.R
\docType{data}
\name{models_GLMNET_MAXENT_SVM}
\alias{models_GLMNET_MAXENT_SVM}
\title{XGBoost Model}
\format{An object of class \code{list} of length 3.}
\usage{
models_GLMNET_MAXENT_SVM
}
\description{
This object...
}
\keyword{datasets}
