% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelSVM.R
\name{modelSVM}
\alias{modelSVM}
\title{modelSVM Function}
\usage{
modelSVM(fileName)
}
\arguments{
\item{fileName}{}
}
\description{
This function allows you to check if a news article is authentic or fake.
It was trained using the SVM Algorithms from the package RTextTools-e1071, and the the Kaggle Fake News Dataset
}
\examples{
modelSVM(FileName)
}
