% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelXGBoost.R
\name{modelXGBoost}
\alias{modelXGBoost}
\title{modelXGBoost Function}
\usage{
modelXGBoost(fileName)
}
\arguments{
\item{fileName}{}
}
\description{
This function allows you to check if a news article is authentic or fake.
It was trained using the xgboost package and the the Kaggle Fake News Dataset
}
\examples{
modelXGBoost(FileName)
}
