% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelXGBoost2.R
\name{modelXGBoost2}
\alias{modelXGBoost2}
\title{modelXGBoost2 Function}
\usage{
modelXGBoost2(fileName)
}
\arguments{
\item{fileName}{}
}
\description{
This function allows you to check if a news article is authentic or fake.
It was trained using the xgboost package and the the Kaggle Fake News Dataset
}
\examples{
modelXGBoost2(FileName)
}
