% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news_article1.R
\docType{data}
\name{news_article1}
\alias{news_article1}
\title{This data can be used to test the modelXGBoost_KFND() function}
\format{An object of class \code{character} of length 1.}
\usage{
news_article1
}
\description{
This object contains the text of a news article...
}
\keyword{datasets}
