% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocab_train.R
\docType{data}
\name{vocab_train}
\alias{vocab_train}
\title{Vocabulary of tokens used to train the XGBoost model}
\format{An object of class \code{text2vec_vocabulary} (inherits from \code{data.frame}) with 354778 rows and 3 columns.}
\usage{
vocab_train
}
\description{
This object...
}
\keyword{datasets}
