% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocab_train_vfn.R
\docType{data}
\name{vocab_train_vfn}
\alias{vocab_train_vfn}
\title{Vocabulary of tokens used to train the XGBoost vfn model}
\format{An object of class \code{text2vec_vocabulary} (inherits from \code{data.frame}) with 203954 rows and 3 columns.}
\usage{
vocab_train_vfn
}
\description{
This object...
}
\keyword{datasets}
